+++ php/Terminus/Models/Backup.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Terminus\Models;
+
+use Terminus\Models\TerminusModel;
+
+class Backup extends TerminusModel {
+}