+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - New command `workflows show` displays details about a workflow (#687)
 - Added back session token-based login. (#693)
 - New command `workflows logs` displays operation logs for workflow (#708)
+- Added initiator data (manual or automated) to `site backups list`. (#716)
 
 ### Fixed
 - `site backups get` no longer errs when there are no backups. (#690)